+++ plom/client/tools/penArrow.py
@@ -153,9 +153,7 @@ class PenArrowItem(QGraphicsItemGroup):
         return ["PenArrow", pth]
 
     def paint(self, painter, option, widget):
-        if not self.collidesWithItem(
-            self.scene().underImage, mode=Qt.ContainsItemShape
-        ):
+        if not self.scene().itemWithinBounds(self):
             # paint a bounding rectangle out-of-bounds warning
             painter.setPen(QPen(QColor(255, 165, 0), 8))
             painter.setBrush(QBrush(QColor(255, 165, 0, 128)))